+++ src/main/webapp/clienti.jsp
@@ -2,7 +2,10 @@
 <%@ page import="com.example.proiectisi.SqlConnection" %>
 <%@ page import="java.sql.PreparedStatement" %>
 <%@ page import="java.sql.ResultSet" %>
-<%@ page import="java.util.Objects" %><%--
+<%@ page import="java.sql.SQLException" %>
+<%@ page import="java.util.List" %>
+<%@ page import="java.util.ArrayList" %>
+<%@ page import="java.util.Arrays" %><%--
   Created by IntelliJ IDEA.
   User: cezar
   Date: 12/28/2021
@@ -17,9 +20,28 @@
 </head>
 <body>
 <%
-    if (!Objects.equals(session.getAttribute("user"), "manager")) {
-        response.sendRedirect("index.jsp");
+    String currentUser = (String) session.getAttribute("user");
+    int codLog = -1;
+    try {
+        Connection connection = SqlConnection.getInstance().getConnection();
+        String sql = "select codf from utilizatori where username = ?;";
+        PreparedStatement stmt = connection.prepareStatement(sql);
+        stmt.setString(1, currentUser);
+        ResultSet rs = stmt.executeQuery();
+
+        if(!rs.next())
+            System.out.println("No Records in the table");
+        else
+            codLog = rs.getInt(1);
+
+    } catch (ClassNotFoundException | SQLException e) {
+        e.printStackTrace();
     }
+
+    List<Integer> allowed = new ArrayList<>(Arrays.asList(4, 6, 7));
+
+    if (!allowed.contains(codLog))
+        response.sendRedirect("index.jsp");
 %>
 
 <div id="nav-placeholder">
@@ -27,9 +49,11 @@
 </div>
 
 <script>
-    $(function(){
+    const cod = <%=codLog %>;
+    if (cod === 4)
         $("#nav-placeholder").load("assets/nav/manager.html");
-    });
+    else if (cod === 6 || cod === 7)
+        $("#nav-placeholder").load("assets/nav/consilier.html");
 </script>
 
 <div id="prod">
@@ -46,10 +70,12 @@
         <input name="judet" type="text" placeholder="Judet">
         <input name="tara" type="text" placeholder="Tara">
 
-        <div class="link">
-            <a class="edit" onclick="exportToExcel('table', 'Clienti')"><img src="${pageContext.request.contextPath}/assets/img/excel.png" alt="Export Excel" title="Export Excel"></a>
-            <a class="edit" onclick="exportToPDF('#table', 'Clienti')"><img src="${pageContext.request.contextPath}/assets/img/pdf.png" alt="Export PDF" title="Export PDF"></a>
-        </div>
+        <% if (codLog != 6) {%>
+            <div class="link">
+                <a class="edit" onclick="exportToExcel('table', 'Clienti')"><img src="${pageContext.request.contextPath}/assets/img/excel.png" alt="Export Excel" title="Export Excel"></a>
+                <a class="edit" onclick="exportToPDF('#table', 'Clienti')"><img src="${pageContext.request.contextPath}/assets/img/pdf.png" alt="Export PDF" title="Export PDF"></a>
+            </div>
+        <% } %>
         <input name="adauga" type="submit" value="Adauga">
     </form>
     <input type='text' id='searchTable' placeholder='Cautare'>
@@ -94,10 +120,12 @@
         <td><%= rs.getString(9)%></td>
         <td><%= rs.getString(10)%></td>
 
-        <td class="link">
-            <a id="edit" href="${pageContext.request.contextPath}/edit/editClienti.jsp?codc=<%= rs.getString(1)%>">Editeaza</a>
-            <a id="delete" href="${pageContext.request.contextPath}/clienti?action=delete&codc=<%= rs.getString(1)%>" methods="">Sterge</a>
-        </td>
+        <% if (codLog != 7) {%>
+            <td class="link">
+                <a id="edit" href="${pageContext.request.contextPath}/edit/editClienti.jsp?codc=<%= rs.getString(1)%>">Editeaza</a>
+                <a id="delete" href="${pageContext.request.contextPath}/clienti?action=delete&codc=<%= rs.getString(1)%>" methods="">Sterge</a>
+            </td>
+        <% } %>
     </tr>
 
     <%} while(rs.next());
